requirements.txt
Django==1.5.1
celery-with-redis==3.0
django-celery==3.0.21
csvkit>=0.5.0
django-debug-toolbar>=0.9.4
django-extensions>=1.1.1
python-dateutil==1.5
python-social-auth>=0.1.2
requests>=1.2.0
requests-oauthlib>=0.3.1
wsgiref==0.1.2
git+git://github.com/tweepy/tweepy.git#egg=tweepy
ipython
# The bleeding edge for rply has iterable token streams
git+git://github.com/alex/rply.git#egg=rply
